+++ Framework/Built_In_Automation/Mobile/Android/adb_calls/adbOptions.py
@@ -373,7 +373,7 @@ def install_app(apk_path, serial=""):
         if serial != "":
             serial = "-s %s" % serial  # Prepare serial number with command line switch
         output = subprocess.check_output(
-            "adb %s install -r %s" % (serial, apk_path), shell=True, encoding="utf-8"
+            "adb %s install -r \"%s\"" % (serial, apk_path), shell=True, encoding="utf-8"
         )
         CommonUtil.ExecLog(sModuleInfo, "Installed app located %s" % apk_path, 0)
         return output