+++ Framework/Built_In_Automation/Sequential_Actions/action_declarations/common.py
@@ -49,7 +49,9 @@ declarations = (
     { "name": "search text",                                 "function": "search_and_save_text",                  "screenshot": "none" },
     { "name": "custom step duration",                        "function": "custom_step_duration",                  "screenshot": "none" },
     { "name": "disable step",                                "function": "disable_step",                          "screenshot": "none" },
-    { "name": "disable showing value",                       "function": "disable_showing_value",                 "screenshot": "none" },          
+    { "name": "disable showing value",                       "function": "disable_showing_value",                 "screenshot": "none" },
+    { "name": "skip testcases",                              "function": "skip_testcases",                        "screenshot": "none" },
+    { "name": "global sleep",                                "function": "global_sleep",                          "screenshot": "none" },
 
     # Mail actions
     { "name": "send mail",                                   "function": "send_mail",                             "screenshot": "none" },